-refactor : 파일 위치 수정 및 item_add_menu_name_menu_image.xml로 파일 이름 수정
